Add VideoRTSPStream class for RTSP video streaming (#57)
* Add VideoRTSPStream class for RTSP video streaming

Introduces the VideoRTSPStream class to manage video capture and streaming from an RTSP source. Supports frame processing, callback registration, configurable FPS, resolution, and JPEG quality, with threading and async event loop support for frame delivery.

* Update docstrings in VideoRTSPStream class

Improved and clarified docstrings for the _start_loop and _capture_loop methods. Updated the description to specify that frames are captured from the RTMP stream instead of the camera.

* Add VideoRTSPStream to video module exports

Imported and added VideoRTSPStream to the __init__.py files in both the root and video submodules, making it available as part of the public API.

* Update default video resolution in VideoRTSPStream

Changed the default resolution from (640, 480) to (480, 680) in both the docstring and the constructor of the VideoRTSPStream class to reflect the new default behavior.

* Fix default video resolution in VideoRTSPStream

Corrects the default resolution from (480, 680) to (480, 640) in both the docstring and the constructor parameter to ensure consistency and accuracy.

* Add AudioRTSPInputStream class and update module exports (#58)

* Add AudioRTSPInputStream class and update module exports

* Fix parameter order in AudioRTSPInputStream constructor

* Add av package dependency to pyproject.toml

* Fix default sampling rate and improve audio chunk handling in AudioRTSPInputStream

* Add audio callback thread to AudioRTSPInputStream for improved processing

* Refactor AudioRTSPInputStream to use multiprocessing for RTSP audio processing and improve control handling

* Refactor logging configuration handling in rtsp_audio_processor and AudioRTSPInputStream

* Improve message sending logic in WebSocket client to handle queue messages more reliably